requirements.txt
# GUI
PySide6>=6.5.0        # または PyQt6>=6.5.0
qt-material>=2.14     # マテリアルデザインテーマ (オプション)

# 画像・動画処理
opencv-python>=4.8.0  # 画像処理
Pillow>=10.0.0        # 画像操作
ffmpeg-python>=0.2.0  # 動画処理
numpy>=1.24.0         # 数値計算
scipy>=1.10.0         # 科学計算

# 機械学習
tensorflow>=2.12.0    # 深層学習フレームワーク
keras>=2.12.0         # 高レベルニューラルネットワークAPI
scikit-learn>=1.2.0   # 機械学習ユーティリティ
onnxruntime>=1.15.0   # ONNX実行環境 (モデル互換性向上)

# データ
SQLAlchemy>=2.0.0     # ORM (オプション)
pydantic>=2.0.0       # データバリデーション
PyYAML>=6.0           # YAML設定ファイル

# 開発ツール
pytest>=7.3.1         # テストフレームワーク
pytest-qt>=4.2.0      # QtテストサポートS
black>=23.3.0         # コードフォーマッター
flake8>=6.0.0         # リンター
mypy>=1.3.0           # 型チェッカー
isort>=5.12.0         # インポート整理

# パッケージング
PyInstaller>=5.10.1   # 実行可能ファイル作成
setuptools>=67.0.0    # パッケージング
wheel>=0.40.0         # ビルドツール